Feat: added single Suggestion component for each suggestions
