+++ src/components/Home/Details/Details.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import Profile from './Profile';
-import Suggestions from './Suggestions';
+import Suggestions from './Suggestions/Suggestions';
 import Footer from './Footer';
 
 const Details = () => {
